requirements.txt
h5py==2.10.0
Keras==2.3.1
keras-tcn==3.1.1
matplotlib==3.3.2
numpy==1.19.5
pandas==1.2.0
scikit-learn==0.23.2
statsmodels==0.12.1
tensorflow-gpu==2.1.0
